kanbanApp/requirements.txt
Django==2.0.2
djangorestframework==3.7.7
pkg-resources==0.0.0
pytz==2018.3
